requirements.txt
# FastAPI and dependencies
fastapi==0.97.0
uvicorn[standard]==0.22.0
websockets==11.0.3
wsproto==1.2.0
prisma==0.9.1
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0
pyjwt==2.10.1

# Hand tracking dependencies
opencv-python==4.8.0.74
numpy==1.24.3
mediapipe==0.10.5
